sui-node: replace validator binary with sui-node
