+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "crates/sui-config",
     "crates/sui-faucet",
     "crates/sui-network",
+    "crates/sui-node",
     "crates/sui-transactional-test-runner",
     "crates/sui-types",
     "crates/sui-verifier",
