+++ crates/sui-node/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "sui-node"
+version = "0.0.0"
+authors = ["Mysten Labs <build@mystenlabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[dependencies]
+anyhow = { version = "1.0.57", features = ["backtrace"] }
+clap = { version = "3.1.17", features = ["derive"] }
+multiaddr = "0.14.0"
+prometheus_exporter = "0.8.4"
+tokio = { version = "1.18.2", features = ["full"] }
+tracing = "0.1.34"
+
+sui-config = { path = "../sui-config" }
+sui_core = { path = "../../sui_core" }
+
+telemetry-subscribers = { git = "https://github.com/MystenLabs/mysten-infra", rev = "7c247967e5a5abd59ecaa75bc62b05bcdf4503fe" }
